// api/src/utilites/achievements.ts
enum Achievements {
    multipleGamesLvl1 = 30,
    multipleGamesLvl2 = 60,
    multipleGamesLvl3 = 90,
    CategoryLvl1 = 30,
    CategoryLvl2 = 60,
    CategoryLvl3 = 90,
    KreditLvl1 = 100,
    KreditLvl2 = 500,
    KreditLvl3 = 999,
    WinLvl1 = 10,
    WinLvl2 = 50,
    WinLvl3 = 99,
    Level5 = 5,
    Level10 = 10,
    Level15 = 15,
}

export default Achievements